1.2.2 Release
# SR
* Added more markers to the TNH Converter
* Fixed the selection box in the Ammo Spawner showing up in the round selection page
* Added Meme tag check on spawned weapons with attachments, system slightly changed and slipped through
* Fixed client enemy multiplier, was loading int data not float
* Random the Random for item spawning
* Started Kiosk

# SRE
* Added Max Patrol Size
* Fixed Opening Levels and Endless Levels
* Added search bar to Library

+++ Packer-SupplyRaid/src/Plugin/src/Scripts/H3MP/SR_Networking.cs
@@ -291,7 +291,7 @@ namespace SupplyRaid
         //Game Options Received
         void GameOptions_Handler(int clientID, Packet packet)
         {
-            float optionPlayerCount = packet.ReadInt();
+            float optionPlayerCount = packet.ReadFloat();
             float optionDifficulty = packet.ReadFloat();
             bool optionFreeBuyMenu = packet.ReadBool();
             bool optionSpawnLocking = packet.ReadBool();
